% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{device_repair}
\alias{device_repair}
\title{Repairability dataset}
\format{
A data frame with 239 rows and 5 variables:
\describe{
\item{device_id}{Unique ID for each device}
\item{repair_time}{Time at which event occurs (unit=days)}
\item{status}{Censoring indicator.  1 = no censoring.  0 = right censoring.  The time value corresponding to status=0 is the out-of-service time for the device}
\item{device_type}{Device SKU (A or B)}
\item{mfg_location}{Manufacturing location (OR, WA, or CA)}
}
}
\usage{
device_repair
}
\description{
Simulated dataset containing event data for 60 devices.
}
\keyword{datasets}
