% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_event.R
\name{geom_event}
\alias{geom_event}
\title{Create event plot}
\usage{
geom_event(
  mapping = NULL,
  ...,
  inherit.aes = T,
  order = T,
  xmax = NULL,
  data = NULL
)
}
\arguments{
\item{mapping}{Aesthetics mapping with the following inputs: \cr
\itemize{
\item time: time column \cr
\item time2: optional secondary time column (e.g. time2 or event/status column specified in Surv() arguments) \cr
\item treatments: A single treatments or a vector of treatments.  Aliased arguments include treatment, trt, and trts. \cr
\item strata: A single stratum or a vector of strata values \cr
}}

\item{...}{Other graphical arguments pertaining to the primary failure/survival curve (color, linetype, linewidth, etc).}

\item{inherit.aes}{Logical, whether to inherit graphics from the ggplot call. TRUE by default.}

\item{order}{Logical, whether to order the event plot based on when it is censored. FALSE by default.}

\item{xmax}{Maximum maximum time for the x-axis.  Defaults to 1.1x maximum of (time1,time2) after the data is converted to interval censored format.}

\item{data}{The data to use, if not provided in the ggplot call}
}
\description{
Create an event plot from provided data
}
\details{
This function creates an event plot from the provided data. On an event plot,
the x-axis is time, while the y-axis has each event separated. By default, the
graph shows solid lines where it is known that an event does not occur, and dashed
lines where an event might have occurred, based on the provided data. If exact
information is known about the event it is marked by an X. Otherwise, the time period
where an event could have occurred is separated by triangles enclosing the dashed area.
}
\examples{
p = mfg \%>\% ggplot()
#Basic plot
p + geom_event(ttr, status)
#Coloring by status
p + geom_event(ttr, status, aes(color = status))
}
