% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_km.R
\name{geom_km}
\alias{geom_km}
\title{Create a Kaplan-Meier Curve}
\usage{
geom_km(
  mapping = NULL,
  ...,
  data = NULL,
  inherit.aes = T,
  conf.int = 0.95,
  failure = F
)
}
\arguments{
\item{mapping}{Aesthetics mapping with the following inputs:
\itemize{
\item time: time column \cr
\item time2: optional secondary time column (e.g. time2 or event/status column specified in Surv() arguments) \cr
\item treatments: optional column or vector corresponding to the treatment variable.  Aliased arguments include treatment, trt, and trts. \cr
\item strata: optional column or vector corresponding to the stratifying variable.   \cr
}}

\item{...}{Other non-aes graphical arguments pertaining to the primary failure/survival curve (color, linetype, linewidth, etc).}

\item{data}{The reference dataset. May be inherited from the plot data or a previous surviveR layer data.}

\item{inherit.aes}{TRUE or FALSE. TRUE inherits arguments from earlier ggplot calls.  Default=TRUE.}

\item{conf.int}{Confidence level for displayed confidence intervals. Must be a value between 0 and 1. Values of FALSE, NULL, or NA will result in no confidence intervals being plotted.}

\item{failure}{TRUE or FALSE (default). Values of TRUE will result in a failure curve being plotted, FALSE will result in a survival curve.}

\item{conf.fill}{Default = NULL, which indicates that confidence intervals will be color-coded by treatment. Other accepted values include the following:
\itemize{
\item String of color name (ex: "black")
\item TRUE indicates the color-coding fill color by treatment.
\item Values of FALSE, NA, and "transparent" may be used to omit shading.
}}

\item{conf.linetype}{Line type of confidence interval lines ("dotted", "dashed", "solid", etc).}

\item{conf.linewidth}{Line width of confidence interval lines.}

\item{conf.color}{Color of confidence interval lines. Value provided must be a string.}

\item{conf.alpha}{Opacity value between 0 and 1.  0=transparent.  1=fully opaque.}
}
\description{
Create a Kaplan-Meier Curve from provided data
}
\details{
This function uses the survfit function in the survival package to generate a
Kaplan-Meier (KM) curve for the given data. Kaplan-Meier curves are a non-parametric
statistic used to estimate the survival function from data. To learn more about the
specifics, see the foundational paper "Nonparametric Estimation from Incomplete Observations" by
E. L. Kaplan and Paul Meier.
}
\examples{

# Univariate survival plot (no treatment or strata variables)
mfg \%>\%
    ggplot() +
    geom_km(aes(time=ttf, time2=status), conf.int = 0.90)


# Univariate failure plot (no treatment or strata variables)
mfg \%>\%
    ggplot() +
    geom_km(aes(time=ttf, time2=status), failure = T)


# Survival plot with treatment variable
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_km(aes(time=ttf, time2=status, treatments = mfg_location), conf.int = 0.90)


# Survival plot with treatment and strata variables
mfg \%>\% ggplot() +
    geom_km(aes(time=ttf, time2=status, treatments = device_type, strata = mfg_location)) +
    facet_grid(.~mfg_location)


# Overriding color & removing confidence intervals (ideal for overlaying fits)
mfg \%>\% ggplot() +
    geom_km(mapping=aes(time=ttf, time2=status, treatments = mfg_location),
            conf.int = F, color="black")


# Overlaying KM curve on semiparametric & parametric fits
mfg \%>\% ggplot() +
    geom_coxph(aes(time=ttf, time2=status, treatments=device_type, strata=mfg_location)) +
    geom_km(color = "black", conf.int = F) +
    facet_grid(.~mfg_location)

mfg \%>\% ggplot() +
    geom_survreg(aes(time=ttf, time2=status, treatments=device_type, strata=mfg_location)) +
    geom_km(color = "black", conf.int = F) +
    facet_grid(.~mfg_location)

mfg \%>\% ggplot() +
    geom_survreg(aes(time=ttf, time2=status, treatments=device_type, strata=mfg_location)) +
    geom_coxph() +
    geom_km(color = "black", conf.int = F) +
    facet_grid(.~mfg_location)

# Note: By default, geom_km2, geom_coxph2, and geom_survreg2 inherit the formula and failure variables.

}
\seealso{
The alternate version that accepts survfit formulas: \link{geom_km2} \cr
The survival function that this function is built on: \link[survival:survfit]{survival::survfit}
}
