% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mcf.R
\name{geom_mcf}
\alias{geom_mcf}
\title{Create Mean Cumulative Function}
\usage{
geom_mcf(mapping = NULL, ..., conf.int = 0.95, data = NULL, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Aesthetics mapping with the following inputs:
\itemize{
\item time: Event times (must be numeric).
\item ids: Unit ID column. Alias arguments include id, Id, Ids, ID, IDS.
\item treatments: A single treatment or a vector of treatments. Alias arguments include treatment, trt, or trts.
\item strata: Grouping variable, typically used for faceting.
}}

\item{...}{Other arguments pertaining to the MCF curves. Examples include, color, linetype, linewidth, etc.}

\item{conf.int}{Confidence level for displayed confidence intervals. Must be a value between 0 and 1. Values of FALSE, NULL, or NA will result in no confidence intervals being plotted.}

\item{data}{The data to use, if not provided in the ggplot call.}

\item{inherit.aes}{TRUE or FALSE. TRUE inherits arguments from earlier ggplot calls.  Default = TRUE.}

\item{conf.fill}{Default = NULL, which indicates that confidence intervals will be color-coded by treatment. Other accepted values include the following:
\itemize{
\item String of color name (ex: "black")
\item TRUE indicates the color-coding fill color by treatment.
\item Values of FALSE, NA, and "transparent" may be used to omit shading.
}}

\item{conf.linetype}{Line type of confidence interval lines ("dotted", "dashed", "solid", etc).}

\item{conf.linewidth}{Line width of the confidence interval lines.}

\item{conf.color}{The line color of confidence interval lines. Value provided must be a string.}

\item{conf.alpha}{The opacity of the confidence interval fill. Values of 0 and 1 correspond to full and no transparency, respectively.}
}
\description{
This function creates a Mean Cumulative Function (MCF) plot from the provided data.
}
\details{
The mean cumulative function (MCF) is a nonparametric survival statistic used to model the average number of events for a repairable
system.  Suppose a system consists of \eqn{N} total system units (hereby termed "units").  Furthermore, let \eqn{i} index the units, and let \eqn{k} index all event times \eqn{t}.
Vector \eqn{(t_1, t_2, ..., t_n)} represents the collective event times of all units in the system.

Each raw observation minimally consists of the following:
\itemize{
\item The unit ID: \eqn{i}
\item The event time: \eqn{t}
\item The risk set status for unit \eqn{i} at time \eqn{t}: \eqn{\delta_i(t)}
where values of 1 indicate that unit \eqn{i} is still active (i.e. in the risk set) and values of 0 indicate that unit \eqn{i} is no longer active (i.e. NOT in the risk set)
\item The event status of unit \eqn{i} at time \eqn{t}: \eqn{d_i(t_k)}
where values of 1 indicate that an event occurred for unit \eqn{i} at time \eqn{t_k} and values of 0 indicate that an event did NOT occur for unit \eqn{i} at time \eqn{t_k}.
}

Letting \eqn{d_.(t_k)} denote the total number of events that happen at time \eqn{t_k} and \eqn{\delta_.(t_k)}
represent the risk set at time \eqn{t_k}, the MCF estimate at time \eqn{t_j} may be expressed as

\deqn{\sum_{k=1}^n \frac{d_.(t_k)}{\delta_.(t_k)} = \sum_{k=1}^n \bar{d}(t_k)}

where \eqn{\bar{d}(t_k)} is the mean cumulative number of events per unit at time \eqn{t_k}.

The confidence intervals geom_mcf displays are Wald estimates, which leverage asymptotic normality assumptions in their calculation.
The nonparametric standard deviation for the MCF at time \eqn{t_k} is estimated using the following expression:

\deqn{\sqrt{ \sum_{i=1}^n \left\{ \sum_{k=1}^j  \frac{\delta_i(t_k)}{\delta_.(t_k)} \times [d_i(t_k) - \bar{d}(t_k)] \right\} ^2}}
Note: The statistical validity (read: bias) of the above MCF statistics hinges on a fairly strong assumption: the recurrence rate of the events
is constant for all units.  This assumption can be violated in common scenarios, so one must check diagnostics to verify the appropriateness of the metric.

Many of the details and notation of this page were heavily adapted from "Statistical Methods
for Reliability Data: Second Edition" (2022) by William Q. Meeker, Luis A. Escobar, and Francis G. Pascual.
For more statistical details on the MCF methodology, please reference pages 45-47 and 522-528 of this fantastic book.
}
\examples{
library(ggplot2); library(dplyr)

# Univariate MCF plot (no treatment or strata variables)
device_repair \%>\% filter(device_type == "A") \%>\%
   ggplot() + 
   geom_mcf(aes(time = repair_time, ids = device_id))


# MCF plot with treatment variable
device_repair \%>\% 
   ggplot() + 
   geom_mcf(aes(time = repair_time, ids = device_id, 
      treatment = mfg_location))
   
   
# MCF plot with strata variable 
device_repair \%>\% 
   ggplot() + 
   geom_mcf(aes(time = repair_time, ids = device_id, 
      strata = device_type))


# MCF plot with treatment and strata variables
device_repair \%>\% 
   ggplot() + 
   geom_mcf(aes(time = repair_time, ids = device_id, 
      treatment = mfg_location, strata = device_type)) +
   facet_grid(.~device_type)

# MCF plot with additional confidence interval arguments
device_repair \%>\% filter(device_type == "A") \%>\%
   ggplot() + 
   geom_mcf(aes(time = repair_time, ids = device_id), 
          conf.int = 0.95, conf.fill = "gray30", 
          conf.linetype = "dotted", conf.linewidth = 0.5, 
          conf.color = "darkred", conf.alpha = 0.15)

}
