% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_survreg.R
\name{geom_survreg2}
\alias{geom_survreg2}
\title{Create a parametric survival plot}
\usage{
geom_survreg2(
  formula = NULL,
  dist = "weibull",
  ...,
  data = NULL,
  inherit.aes = T,
  conf.int = 0.95,
  failure = F,
  length.out = 1000,
  xmax = NULL
)
}
\arguments{
\item{formula}{Survival formula}

\item{dist}{Distribution to be used in parametric survival fit.  Default value: "weibull".  Accepts all distributions accepted by \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}, including
"genf","genf.orig" ,"gengamma", "gengamma.orig", "exp", "weibull", "weibullPH", "lnorm", "gamma", "gompertz", "llogis", "exponential", and "lognormal".  For a list of distributions supported by this function, please reference the documentation for \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}.}

\item{...}{Other graphical arguments pertaining to the primary failure/survival curve (color, linetype, linewidth, etc).}

\item{data}{The reference dataset. May be inherited from the plot data or a previous surviveR layer data.}

\item{inherit.aes}{TRUE or FALSE. TRUE inherits arguments from earlier ggplot calls.  Default=TRUE.}

\item{conf.int}{Confidence level for displayed confidence intervals. Must be a value between 0 and 1. Values of FALSE, NULL, or NA will result in no confidence intervals being plotted.}

\item{failure}{TRUE or FALSE (default).  Values of TRUE will result in a failure curve being plotted.}

\item{length.out}{length of the time vector created in the prediction dataset that is plotted}

\item{xmax}{Maximum time value on x-axis.  Defaults to 1.15 times max(time) if exact or right-censored data and 1.15 times max(time2) for interval-censored data.}

\item{conf.fill}{Default = NULL, which indicates that confidence intervals will be color-coded by treatment. Other accepted values include the following:
\itemize{
\item String of color name (ex: "black")
\item TRUE indicates the color-coding fill color by treatment.
\item Values of FALSE, NA, and "transparent" may be used to omit shading.
}}

\item{conf.linetype}{Line type of confidence interval lines ("dotted", "dashed", "solid", etc).}

\item{conf.linewidth}{Line width of confidence interval lines.}

\item{conf.color}{Color of confidence interval lines. Value provided must be a string.}

\item{conf.alpha}{Opacity value between 0 and 1.  0=transparent.  1=fully opaque.}
}
\description{
Create a parametric survival plot using formula-based arguments
}
\details{
This function allows users to visualize parametric survival fits using
the ggplot Grammar of Graphics.  It leverages the flexsurvreg function in the \link[flexsurv:flexsurv-package]{flexsurv::flexsurv-package}
package.  For more information on the distributions supported by geom_survreg and
geom_survreg2, please see the documentation for \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}.  Note that
flexsurvreg (and, by extension, geom_survreg2) may run sluggishly with large datasets.
}
\examples{
# Parametric survival plot with treatment variable (default: Weibull distribution)
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_survreg2(formula = Surv(ttf,status) ~ mfg_location, conf.int = 0.90)

# ^ equivalent to
# mfg \%>\% filter(device_type == "A") \%>\%
#      ggplot() +
#      geom_survreg(aes(time=ttf, time2=status, treatments = mfg_location), conf.int = 0.90)


# Lognormal survival fit with treatment variable
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_survreg2(formula = Surv(ttf,status) ~ mfg_location, dist = "lognormal")

# ^ equivalent to
# mfg \%>\% filter(device_type == "A") \%>\%
#      ggplot() +
#      geom_survreg(aes(time=ttf, time2=status, treatments = mfg_location), dist = "lognormal")


# Parametric failure plot with treatment variable and 90\% confidence interval
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_survreg2(formula = Surv(ttf,status) ~ mfg_location, failure = T, conf.int = 0.90)

# ^ equivalent to
# mfg \%>\% filter(device_type == "A") \%>\%
#      ggplot() +
#      geom_survreg(aes(time=ttf, time2=status, treatments = mfg_location), failure = T, conf.int = 0.90)


# Weibull fit with treatment and strata variables
mfg \%>\% ggplot() +
    geom_survreg2(formula = Surv(ttf,status) ~ treatments + strata(mfg_location))
    facet_grid(.~mfg_location)

# ^ equivalent to
# mfg \%>\% ggplot() +
#      geom_survreg(aes(time=ttf, time2=status, treatments = device_type, strata = mfg_location)) +
#      facet_grid(.~mfg_location)


# Overlaying KM curve on top of parametric fit
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_survreg2(formula = Surv(ttf,status) ~ mfg_location)) +
    geom_km2(conf.int = F)

# Note: By default, geom_km2, geom_coxph2, and geom_survreg2 inherit the formula and failure variables.
}
