% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_add.survivalverse_plot.R,
%   R/survivalverse_general_functions.R
\name{ggplot_add.survivalverse_plot}
\alias{ggplot_add.survivalverse_plot}
\title{ggplot_add}
\usage{
\method{ggplot_add}{survivalverse_plot}(object, plot, object_name)

\method{ggplot_add}{survivalverse_plot}(object, plot, object_name)
}
\description{
ggplot_add

ggplot_add
}
\details{
ggplot_add

ggplot_add
}
