% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surviveR.R
\docType{package}
\name{surviveR}
\alias{surviveR-package}
\alias{surviveR}
\title{surviveR: Tidyverse-based Reliability and Survival Data Visualization Tool}
\description{
A suite of reliability/survival data visualization tools that leverage
the Tidyverse Grammar of Graphics.  Time-based aesthetics may be specified
alongside standard non-aesthetic arguments to generate highly customizable
graphs.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KLGORE/surviveR}
  \item \url{https://klgore.github.io/surviveR/}
  \item Report bugs at \url{https://github.com/KLGORE/surviveR/issues}
}

}
\author{
Kristen Gore, Ph.D., Sam Jeffe, Emma Schneider
}
