% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_coxph.R
\name{geom_coxph}
\alias{geom_coxph}
\title{Fit a Cox proportional Hazards Regression Model}
\usage{
geom_coxph(
  mapping = NULL,
  ...,
  conf.int = 0.95,
  failure = FALSE,
  data = NULL,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Aesthetics mapping with the following inputs:
\itemize{
\item time: time column \cr
\item time2: optional secondary time column (e.g. time2 or event/status column specified in Surv() arguments) \cr
\item treatments: optional column or vector corresponding to the treatment variable.  Aliased arguments include treatment, trt, and trts. \cr
\item strata: optional column or vector corresponding to the stratifying variable.   \cr
}}

\item{...}{Other non-aes graphical arguments pertaining to the primary failure/survival curve (color, linetype, linewidth, etc).}

\item{conf.int}{Confidence level for displayed confidence intervals. Must be a value between 0 and 1. Values of FALSE, NULL, or NA will result in no confidence intervals being plotted.}

\item{failure}{TRUE or FALSE (default).  Values of TRUE will result in a failure curve being plotted.}

\item{data}{The reference dataset. May be inherited from the plot data or a previous surviveR layer data.}

\item{inherit.aes}{TRUE or FALSE. TRUE inherits arguments from earlier ggplot calls.  Default=TRUE.}

\item{conf.fill}{Default = NULL, which indicates that confidence intervals will be color-coded by treatment. Other accepted values include the following:
\itemize{
\item String of color name (ex: "black")
\item TRUE indicates the color-coding fill color by treatment.
\item Values of FALSE, NA, and "transparent" may be used to omit shading.
}}

\item{conf.linetype}{Line type of confidence interval lines ("dotted", "dashed", "solid", etc).}

\item{conf.linewidth}{Line width of confidence interval lines.}

\item{conf.color}{Color of confidence interval lines. Value provided must be a string.}

\item{conf.alpha}{Opacity value between 0 and 1.  0 = transparent.  1 = fully opaque.}
}
\description{
Utilizes tidyverse syntax to visualize Cox proportional hazards predictions.
}
\details{
This function uses the coxph function in the survival package to visualize Cox Proportional Hazards (CPH) survival curves within the Tidyverse ecosystem.
The Cox proportional hazards regression model is a versatile semiparametric method used to model survival data.  It applies to cases in which hazards can be modeled as follows:

\deqn{ h(t) = h_0(t) \exp (b_1 x_1 + b_2 x_2 + \ldots )}

upon estimation of the model coefficients, the above formula can be reexpressed in terms of the survival probability.  This expression is given by

\deqn{ S(t) = S_0(t)^{\exp(\beta)} }
\deqn{ \log[S(t)] = \exp(\beta) \log[S_0(t)] }

Full details of this model are found in the foundational "Regression Models and Life-Tables" paper by D.R. Cox (1972).
Note: Though this function accepts strata in its arguments, the CPH predictions are identical to that of the model in which the stratifying variable is a treatment instead.  Specifying strata only impacts the line type.
}
\examples{
library(ggplot2); library(dplyr)

# CoxPH survival plot with treatment variable
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_coxph(aes(time = ttf, time2 = status, 
        treatments = mfg_location), conf.int = 0.90)


# CoxPH failure plot with treatment variable
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_coxph(aes(time = ttf, time2 = status, 
        treatments = mfg_location), failure = TRUE)


# CoxPH survival plot with treatment and strata variables
mfg \%>\% 
    ggplot() +
    geom_coxph(aes(time = ttf, time2 = status, 
        treatments = device_type, strata = mfg_location)) +
    facet_grid(.~mfg_location)


# Overlaying KM curve on CoxPH fit
mfg \%>\% 
    ggplot() +
    geom_coxph(aes(time = ttf, time2 = status, 
        treatments = device_type, strata = mfg_location)) +
    geom_km(color = "black", conf.int = FALSE) +
    facet_grid(.~mfg_location)

# Note: By default, geom_km, geom_coxph, and geom_survreg 
# inherit the formula and failure variables.


}
\seealso{
Note: the \link{geom_coxph2} function allows for an input of a survival formula rather than specifying elements through the aesthetics mapping. \cr
}
