% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_km.R
\name{geom_km2}
\alias{geom_km2}
\title{Create a Kaplan-Meier Curve Using the Tidyverse Grammar-of-Graphics}
\usage{
geom_km2(
  formula = NULL,
  ...,
  conf.int = 0.95,
  failure = FALSE,
  data = NULL,
  inherit.aes = T
)
}
\arguments{
\item{formula}{A formula compatible with survfit(), or a survfit object}

\item{...}{Non-aes graphical arguments pertaining to the primary failure/survival curve (color, linetype, linewidth, etc).}

\item{conf.int}{Confidence level for displayed confidence intervals. Must be a value between 0 and 1. Values of FALSE, NULL, or NA will result in no confidence intervals being plotted.}

\item{failure}{TRUE or FALSE (default). Values of TRUE will result in a failure curve being plotted, FALSE will result in a survival curve.}

\item{data}{The reference dataset. May be inherited from the plot data or a previous surviveR layer data.}

\item{inherit.aes}{TRUE or FALSE. TRUE inherits arguments from earlier ggplot calls.  Default = TRUE.}
}
\description{
Create Kaplan-Meier curve by providing a survival formula
}
\details{
The \code{geom_km2} function generates KM curves that display nonparametric
survival estimates S(t) versus time (t).  Unlike its companion function \code{geom_km},
\code{geom_km2} allows users to input a survival formula instead of specifying
the aesthetics mapping.  Cumulative failure curves can be plotting
in lieu of the survival probability by using the \code{failure=TRUE} argument.  For more
statistical details on the Kaplan-Meier methodology, please see chapter 3 of
\emph{Statistical Methods for Reliability Data (2nd edition)} by Meeker, Escobar,
and Pascual.

\bold{Note on additional non-aesthetics arguments:} \cr
Users may customize graphical attributes of the primary KM curve(s)
and confidence intervals.  Options include the following:
\itemize{
\item KM Curve(s): color, linewidth, linetype, alpha \cr
\item Confidence Interval(s): conf.color, conf.fill, conf.linewidth, conf.linetype, conf.alpha  \cr
\itemize{
\item \emph{conf.color}: Color of confidence interval lines. Value provided must be a string.
\item \emph{conf.fill}: Default = NULL, which indicates that confidence intervals will be
color-coded by treatment. Other accepted values include the following:
\itemize{
\item String of color name (ex: "black")
\item TRUE indicates the color-coding fill color by treatment.
\item Values of FALSE, NA, and "transparent" may be used to omit shading.
}
\item \emph{conf.linewidth}: Line width of confidence interval lines.
\item \emph{conf.linetype}: Line type of confidence interval lines ("dotted", "dashed", "solid", etc).
\item \emph{conf.alpha}: Opacity value between 0 and 1.  0 = transparent.  1 = fully opaque.
}
}
}
\examples{
library(ggplot2); library(dplyr)

# Univariate survival plot (no treatment or strata variables)
mfg \%>\%
    ggplot() +
    geom_km2(formula = Surv(ttf,status) ~ 1, conf.int = 0.90)

# ^ equivalent to
# mfg \%>\%
#      ggplot() +
#      geom_km(aes(time = ttf, time2 = status), 
#           conf.int = 0.90)


# Univariate failure plot (no treatment or strata variables)
mfg \%>\%
    ggplot() +
    geom_km2(
        formula = Surv(ttf,status) ~ 1, 
        failure = TRUE)

# ^ equivalent to
# mfg \%>\%
#      ggplot() +
#      geom_km(aes(time = ttf, time2 = status), failure = TRUE)


# Survival plot with treatment variable
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_km2(
        formula = Surv(ttf,status) ~ mfg_location, 
        conf.int = 0.90)

# ^ equivalent to
# mfg \%>\% filter(device_type == "A") \%>\%
#      ggplot() +
#      geom_km(aes(time = ttf, time2 = status, 
#           treatments = mfg_location), conf.int = 0.90)


# Survival plot with treatment and strata variables
mfg \%>\% 
    ggplot() +
    geom_km2(
        formula = Surv(ttf, status) ~ device_type + 
                  strata(mfg_location)) +
    facet_grid(.~mfg_location)

# ^ equivalent to
# mfg \%>\% 
#      ggplot() +
#      geom_km(aes(time = ttf, time2 = status, 
#           treatments = device_type, strata = mfg_location)) +
#      facet_grid(.~mfg_location)


# Overlaying KM curve on semiparametric & parametric fits
mfg \%>\% 
    ggplot() +
    geom_coxph2(formula = Surv(ttf, status) ~ 
                  device_type + strata(mfg_location)) +
    geom_km2(color = "black", conf.int = FALSE) +
    facet_grid(.~mfg_location)

mfg \%>\% 
    ggplot() +
    geom_survreg2(formula = Surv(ttf, status) ~ 
        device_type + strata(mfg_location)) +
    geom_km2(color = "black", conf.int = FALSE) +
    facet_grid(.~mfg_location)

mfg \%>\% 
    ggplot() +
    geom_survreg2(formula = Surv(ttf, status) ~ 
        device_type + strata(mfg_location)) +
    geom_coxph2() +
    geom_km2(color = "black", conf.int = FALSE) +
    facet_grid(.~mfg_location)
# Note: By default, geom_km2, geom_coxph2, and 
# geom_survreg2 inherit the formula and failure variables.

}
\seealso{
\link{geom_km}. \cr
}
