% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mcf.R
\name{geom_mcf}
\alias{geom_mcf}
\title{Create Mean Cumulative Function}
\usage{
geom_mcf(mapping = NULL, ..., data = NULL, inherit.aes = T)
}
\arguments{
\item{mapping}{Aesthetics mapping with the following inputs:
\itemize{
\item time: Event times (must be numeric).
\item ids: Unit ID column. Alias arguments include id, Id, Ids, ID, IDS.
\item treatments: A single treatment or a vector of treatments. Alias arguments include treatment, trt, or trts.
\item strata: Grouping variable, typically used for faceting.
}}

\item{...}{Other arguments pertaining to the MCF curves. Examples include, color, linetype, linewidth, etc.}

\item{data}{The reference dataset.}

\item{inherit.aes}{TRUE or FALSE. TRUE inherits arguments from earlier ggplot calls.  Default=TRUE.}

\item{conf.int}{Confidence level for displayed confidence intervals. Must be a value between 0 and 1. Values of FALSE, NULL, or NA will result in no confidence intervals being plotted.}

\item{conf.fill}{Default = NULL, which indicates that confidence intervals will be color-coded by treatment. Other accepted values include the following:
\itemize{
\item String of color name (ex: "black")
\item TRUE indicates the color-coding fill color by treatment.
\item Values of FALSE, NA, and "transparent" may be used to omit shading.
}}

\item{conf.linetype}{Line type of confidence interval lines ("dotted", "dashed", "solid", etc).}

\item{conf.linewidth}{Line width of the confidence interval lines.}

\item{conf.color}{The line color of confidence interval lines. Value provided must be a string.}

\item{conf.alpha}{The opacity of the confidence interval fill. Values of 0 and 1 correspond to full and no transparency, respectively.}
}
\description{
This function creates a Mean Cumulative Function (MCF) plot from the provided data.
}
\details{
This function uses the mean cumulative function to create a nonparametric model fit within the Tidyverse ecosystem.
The mean cumulative incidence function (MCF) is a nonparametric display for reparable systems. The MCF estimate is
nonparametric because it has minimal assumptions and does not use a parametric model for the population MCF. The
basic assumption of this estimate is that a population of cumulative functions exists, with one for each system in
the population. It is also required that a system's observation time is not dependent on the history of the system.
This would result in biased MCF estimators. The cumulative incidence function from Chapter 22 in "Statistical Methods
for Reliability Data: Second Edition" (2022) by William Q. Meeker, Luis A. Escobar, and Francis G. Pascual is used in
the creation of this function. The aesthetics of this function match those of the tidyverse.

The following Tidyverse aesthetics arguments are accepted in geom_mcf: color, fill, group, linetype, and linewidth.
}
\examples{
Example 1: Single device

  set.seed(762)
  simulating data for a single device
  df=tibble(
      device_type="A",
      time=rpois(100,lambda=5),
      event_time=cumsum(time)
  )

  df \%>\% ggplot() + geom_mcf(mapping = aes(time=event_time))

# Example 2: device_repair (multistrata example) -> trts instead?

p = device_repair \%>\% ggplot()

# Basic mcf plot, providing ID and strata    is this basic??? #When ID and strata are provided. 
p + geom_mcf(mapping = aes(time=repair_time, ids=device_id, strata=device_type))

# When no strata is provided, #When ID is provided, but no strata.
p + geom_mcf(mapping = aes(time=repair_time, ids=device_id))

# If no ID is provided a generic one will be there #When strata is provided, but no ID.
p + geom_mcf(mapping = aes(time=repair_time, strata=device_type))

# should we have a color = strata?

# All the confidence interval arguments provided
# filtered ds for this one to just show A or something?
p + geom_mcf(mapping = aes(time=repair_time, ids=device_id), conf.int=0.95,
             conf.fill = "gray30", conf.linetype = "dotted",
             conf.width = 0.5, conf.color = "darkred", conf.alpha = 0.15)


}
