% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_survreg.R
\name{geom_survreg}
\alias{geom_survreg}
\title{Create a Parametric Survival Curve Using the Tidyverse Grammar-of-Graphics}
\usage{
geom_survreg(
  mapping = NULL,
  dist = "weibull",
  ...,
  conf.int = 0.95,
  failure = FALSE,
  length.out = 1000,
  xmax = NULL,
  data = NULL,
  inherit.aes = T
)
}
\arguments{
\item{mapping}{Aesthetics mapping with the following inputs: \cr
\itemize{
\item time: time column \cr
\item time2: optional secondary time column (e.g. time2 or event/status column specified in Surv() arguments) \cr
\item treatments: optional column or vector corresponding to the treatment variable.  Aliased arguments include treatment, trt, and trts. \cr
\item strata: optional column or vector corresponding to the stratifying variable.   \cr
}}

\item{dist}{Distribution to be used in parametric survival fit.  Default value: "weibull".
Accepts all distributions accepted by \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}, including
"genf","genf.orig" ,"gengamma", "gengamma.orig", "exp", "weibull", "weibullPH", "lnorm", "gamma", "gompertz", "llogis", "exponential", and "lognormal".  For a list of distributions supported by this function, please reference the documentation for \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}.}

\item{...}{Non-aes graphical arguments pertaining to the primary failure/survival curve (color, linetype, linewidth, etc).}

\item{conf.int}{Confidence level for displayed confidence intervals. Must be a value between 0 and 1. Values of FALSE, NULL, or NA will result in no confidence intervals being plotted.}

\item{failure}{TRUE or FALSE (default).  Values of TRUE will result in a failure curve being plotted.}

\item{length.out}{Length of the time vector used in the prediction dataset being plotted.}

\item{xmax}{Maximum time value on x-axis.  Defaults to 1.15 times max(time) if exact or right-censored data and 1.15 times max(time2) for interval-censored data.}

\item{data}{The reference dataset. May be inherited from the plot data or a previous surviveR layer data.}

\item{inherit.aes}{TRUE or FALSE. TRUE inherits arguments from earlier ggplot calls.  Default = TRUE.}
}
\description{
Create a parametric survival curve using aesthetic mapping arguments
}
\details{
The \code{geom_survreg} function generates parametric survival curves for survival S(t) versus time (t).
It leverages the flexsurvreg function in the \link[flexsurv:flexsurv-package]{flexsurv::flexsurv-package} package.  For more information on the
distributions supported by geom_survreg and geom_survreg2, please see the
documentation for \link[flexsurv:flexsurvreg]{flexsurv::flexsurvreg}.

Note that \code{geom_survreg} (and, by extension, \code{geom_survreg2}) may run sluggishly
with large datasets.

\bold{Note on additional non-aesthetics arguments:} \cr
Users may customize graphical attributes of the primary parametric curve(s)
and confidence intervals.  Options include the following:
\itemize{
\item Parametric Curve(s): color, linewidth, linetype, alpha \cr
\item Confidence Interval(s): conf.color, conf.fill, conf.linewidth, conf.linetype, conf.alpha  \cr
\itemize{
\item \emph{conf.color}: Color of confidence interval lines. Value provided must be a string.
\item \emph{conf.fill}: Default = NULL, which indicates that confidence intervals will be
color-coded by treatment. Other accepted values include the following:
\itemize{
\item String of color name (ex: "black")
\item TRUE indicates the color-coding fill color by treatment.
\item Values of FALSE, NA, and "transparent" may be used to omit shading.
}
\item \emph{conf.linewidth}: Line width of confidence interval lines.
\item \emph{conf.linetype}: Line type of confidence interval lines ("dotted", "dashed", "solid", etc).
\item \emph{conf.alpha}: Opacity value between 0 and 1.  0 = transparent.  1 = fully opaque.
}
}
}
\examples{
library(ggplot2); library(dplyr)

# Parametric survival plot with treatment variable (default: Weibull distribution)
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_survreg(aes(time = ttf, time2 = status, 
        treatments = mfg_location), conf.int = 0.90)


# Lognormal survival fit with treatment variable
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_survreg(aes(time = ttf, time2 = status, 
        treatments = mfg_location), dist = "lognormal")


# Parametric failure plot with treatment variable
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_survreg(aes(time = ttf, time2 = status, 
        treatments = mfg_location), failure = TRUE, 
        conf.int = 0.90)


# Weibull fit with treatment and strata variables
mfg \%>\% 
    ggplot() +
    geom_survreg(aes(time = ttf, time2 = status, 
        treatments = device_type, strata = mfg_location)) +
    facet_grid(.~mfg_location)


# Overlaying KM curve on top of parametric fit
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_survreg(aes(time = ttf, time2 = status, 
        treatments = mfg_location)) +
    geom_km(conf.int = FALSE)

# Note: By default, geom_km, geom_coxph, and 
# geom_survreg inherit the mapping (time, time2, treatments, strata) and failure variables.
}
\seealso{
\link{geom_survreg2}, which accepts a survival formula in lieu of variable
specification using aesthetics mappings. \cr
}
