% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_weibull.R
\name{geom_weibull}
\alias{geom_weibull}
\title{Create a Weibull Plot}
\usage{
geom_weibull(mapping = NULL, ..., data = NULL, inherit.aes = T, line.fit = T)
}
\arguments{
\item{mapping}{Aesthetics mapping with the following inputs: \cr
\itemize{
\item time: time column \cr
\item time2: optional secondary time column (e.g. time2 or event/status column specified in Surv() arguments) \cr
\item treatments: optional column or vector corresponding to the treatment variable.  Aliased arguments include treatment, trt, and trts \cr
\item strata: optional column or vector corresponding to the stratifying variable  \cr
}}

\item{...}{Other graphical arguments pertaining to the primary failure/survival curve (color, linetype, linewidth, etc) \cr}

\item{data}{The reference dataset. May be inherited from the plot data or a previous surviveR layer data.}

\item{conf.int}{Confidence level for displayed confidence intervals. Must be a value between 0 and 1. Values of FALSE, NULL, or NA will result in no confidence intervals being plotted.}

\item{conf.fill}{Default = NULL, which indicates that confidence intervals will be color-coded by treatment. Other accepted values include the following:
\itemize{
\item String of color name (ex: "black") \cr
\item TRUE indicates the color-coding fill color by treatment. \cr
\item Values of FALSE, NA, and "transparent" may be used to omit shading. \cr
}}

\item{conf.linetype}{Line type of confidence interval lines ("dotted", "dashed", "solid", etc).}

\item{conf.linewidth}{Line width of confidence interval lines.}

\item{conf.color}{Color of confidence interval lines. Value provided must be a string.}

\item{conf.alpha}{Opacity value between 0 and 1.  0=transparent.  1=fully opaque.}
}
\description{
Create a Weibull plot using Kaplan-Meier tables
}
\details{
This function uses the survfit function in the survival package to generate a
Weibull plot for the given data.  Weibull plots showcase log(-log(S(t)) vs. time relationship.
Data that is well-modeled by the Weibull distribution should yield a linear relationship.  ADD REFERENCE

The \link{geom_weibull} function color-codes data points by the treatment variable and assigns shape according to the strata variable.
}
\examples{
# Univariate Weibull plot (no treatment or strata variables)
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_weibull(aes(time=ttf, time2=status))


# Weibull plot with treatment variable
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_weibull(aes(time=ttf, time2=status, treatments = mfg_location))


# Weibull plot with no line fit
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_weibull(aes(time=ttf, time2=status, treatments = mfg_location), line.fit = F)


# Weibull plot with treatment and strata variables
mfg \%>\% ggplot() +
    geom_weibull(aes(time=ttf, time2=status, treatments = device_type, strata = mfg_location)) +
    facet_grid(.~mfg_location, scale="free_x")

}
