% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_weibull.R
\name{geom_weibull}
\alias{geom_weibull}
\title{Create a Weibull Plot}
\usage{
geom_weibull(
  mapping = NULL,
  ...,
  line.fit = TRUE,
  data = NULL,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping}{Aesthetics mapping with the following inputs: \cr
\itemize{
\item time: time column \cr
\item time2: optional secondary time column (e.g. time2 or event/status column specified in Surv() arguments) \cr
\item treatments: optional column or vector corresponding to the treatment variable.  Aliased arguments include treatment, trt, and trts \cr
\item strata: optional column or vector corresponding to the stratifying variable  \cr
}}

\item{...}{Other graphical arguments pertaining to the primary failure/survival curve (color, linetype, linewidth, etc) \cr}

\item{line.fit}{TRUE or FALSE.  Indicates whether a line should be plotted for each subgroup.  Default = TRUE.}

\item{data}{The reference dataset. May be inherited from the plot data or a previous surviveR layer data.}

\item{inherit.aes}{TRUE or FALSE. TRUE inherits arguments from earlier ggplot calls.  Default = TRUE.}
}
\description{
Create a Weibull Plot Using the Tidyverse Grammar-of-Graphics

The \code{geom_weibull} function utilizes \link[survival:survfit]{survival::survfit} to generate a
Weibull plot for the provided survival data.  As discussed in chapter 6 of
\emph{Statistical Methods for Reliability Data (2nd edition)} by Meeker, Escobar,
and Pascual (2022), data that is well-modeled by the Weibull distribution should yield a linear
log(-log(S(t)) vs. log(t) relationship, where \eqn{t} represents time and
\eqn{S(t)} represents the nonparametric survival probability estimate.

The \link{geom_weibull} function color-codes data points by the
treatment variable.  It assigns data point characters
according to the strata variable.

\bold{Note on additional non-aesthetics arguments:} \cr
Users may customize graphical attributes of the data points
and corresponding line fit(s).  Options include the following:
\itemize{
\item Weibull Points: color, size, alpha \cr
\item Line fit: line.color, linewidth, linetype, line.alpha  \cr
\itemize{
\item \emph{line.color}: Color of  fit line(s). Must be a string.
\item \emph{linewidth}: Line width of fit line(s).
\item \emph{linetype}: Line type of the fit line(s) ("dotted", "dashed", "solid", etc).
\item \emph{line.alpha}: Opacity of fit line(s).  0 = transparent.  1 = fully opaque.
}
}
}
\examples{
library(ggplot2); library(dplyr)

# Univariate Weibull plot (no treatment or strata variables)
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_weibull(aes(time = ttf, time2 = status))


# Weibull plot with treatment variable
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_weibull(aes(time = ttf, time2 = status, 
        treatments = mfg_location))


# Weibull plot with no line fit
mfg \%>\% filter(device_type == "A") \%>\%
    ggplot() +
    geom_weibull(aes(time = ttf, time2 = status, 
        treatments = mfg_location), line.fit = FALSE)


# Weibull plot with treatment and strata variables
mfg \%>\% ggplot() +
    geom_weibull(aes(time = ttf, time2 = status, 
        treatments = device_type, strata = mfg_location)) +
    facet_grid(.~mfg_location, scale = "free_x")

}
